analogy wip
